refactor[#4] : 깃 크라켄 관련 문제 해결
- 깃 크라켄 사용하며 파일들이 자동 커밋으로 올라가는 문제들이 발생했어서 해당 문제를 해결하기 위해 일단 모든 파일을 커밋하고 푸시합니다.
