+++ backend/src/main/java/com/handong/rebon/menu/domain/Menu.java
@@ -2,7 +2,7 @@ package com.handong.rebon.menu.domain;
 
 import javax.persistence.*;
 
-import com.handong.rebon.shop.domain.item.Shop;
+import com.handong.rebon.member.domain.item.Shop;
 
 @Entity
 public class Menu {
