+++ backend/src/main/java/com/handong/rebon/review/domain/Review.java
@@ -3,12 +3,12 @@ package com.handong.rebon.review.domain;
 import java.util.List;
 import javax.persistence.*;
 
-import com.handong.rebon.member.domain.Member;
+import com.handong.rebon.member.Member;
 import com.handong.rebon.review.domain.content.ReviewContent;
 import com.handong.rebon.review.domain.content.ReviewImages;
 import com.handong.rebon.review.domain.content.ReviewScore;
 import com.handong.rebon.review.domain.empathy.Empathy;
-import com.handong.rebon.shop.domain.item.Shop;
+import com.handong.rebon.member.domain.item.Shop;
 
 @Entity
 public class Review {
