+++ backend/src/main/java/com/handong/rebon/review/domain/empathy/Empathy.java
@@ -2,7 +2,7 @@ package com.handong.rebon.review.domain.empathy;
 
 import javax.persistence.*;
 
-import com.handong.rebon.member.domain.Member;
+import com.handong.rebon.member.Member;
 import com.handong.rebon.review.domain.Review;
 
 @Entity
